feat(custompill): CLIN-1934 fix review comments
