+++ packages/ui/src/components/CustomPill/QueriesSidebar/EditCustomPillModal.module.scss
@@ -1,7 +1,7 @@
 @import 'theme.template';
 
 .modalContainer {
-    width: 100% !important;
+    width: 100%;
     max-width: 100%;
     margin: 0;
     top: 0;
@@ -17,7 +17,7 @@
                 color: white;
 
                 span[class$='-title-content'] {
-                    margin-left: 0 !important;
+                    margin-left: 0;
                 }
             }
         }
